trustcontrol_extracted/trustcontrol/requirements.txt
# ════════════════════════════════════════════════════════════
#  TrustControl — Зависимости сервера
#  pip install -r requirements.txt
# ════════════════════════════════════════════════════════════

# Сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0

# База данных
sqlalchemy==2.0.25
aiosqlite==0.19.0          # SQLite async (для старта)
asyncpg==0.29.0            # PostgreSQL async (для продакшна)
alembic==1.13.1            # миграции БД

# Авторизация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# OpenAI / Anthropic
openai==1.10.0
anthropic==0.18.0

# Telegram
python-telegram-bot==20.7

# HTTP запросы
httpx==0.26.0
requests==2.31.0

# Валидация данных
pydantic[email]==2.5.3

# ════════════════════════════════════════════════════════════
#  Зависимости скрипта на кассе (monitor.py)
#  pip install -r requirements-monitor.txt
# ════════════════════════════════════════════════════════════
